Conflicts:
	intlibs/dll/dll.vcproj
	runtimes/cpp/platforms/sdl/MoRE/MoRE.vcproj
	runtimes/cpp/platforms/sdl/mosynclib/mosynclib.vcproj
	runtimes/cpp/platforms/sdl/sdl.vcproj
	tools/mhsd-display/mhsd-display.vcproj

+++ tests/libc-suite/setup_filesystem.c
@@ -108,7 +108,7 @@ static bool makeDir(char* root, int rootLen, const char* name, int bufLen) {
 		int resLen = rootLen + res;
 		if(root[resLen-1] == '/') {
 			rootLen = resLen;
-			MAASSERT(rootLen + strlen(name) < bufLen);
+			MAASSERT(rootLen + (int)strlen(name) < bufLen);
 			//printf("Dir: '%s'\n", file.c_str());
 			strcpy(root + rootLen, name);
 			if(tryToMake(root))
